Add Unhealthy Error Text
If there is not an exception or a description on the HealthReportEntry,
build an error text which reports the name and Status.

The
[Microsoft.Extensions.Diagnostics.HealthChecks.EntityFrameworkCore](https://www.nuget.org/packages/Microsoft.Extensions.Diagnostics.HealthChecks.EntityFrameworkCore)
sets neither the description nor sets an exception in some cases (i.e.
the database if offline). This causes a confusing message from PRTG
where it reports an error but then uses the default text "Everything is
fine :)".